Use int instead of long

+++ src/main/java/me/webhead1104/township/commands/CashCommand.java
@@ -7,13 +7,13 @@ import org.bukkit.entity.Player;
 
 public final class CashCommand {
 
-    public static void addCash(Player player, long cash) {
+    public static void addCash(Player player, int cash) {
         User user = Township.getUserManager().getUser(player.getUniqueId());
         user.setCash(user.getCash() + cash);
         player.sendMessage(Msg.format("<green>Added %d cash!", cash));
     }
 
-    public static void removeCash(Player player, long cash) {
+    public static void removeCash(Player player, int cash) {
         User user = Township.getUserManager().getUser(player.getUniqueId());
         if (user.getCash() - cash < 0) {
             player.sendMessage(Msg.format("<red>Unable to remove %s cash!", cash));
@@ -23,7 +23,7 @@ public final class CashCommand {
         player.sendMessage(Msg.format("<green>Removed %s cash!", cash));
     }
 
-    public static void setCash(Player player, long cash) {
+    public static void setCash(Player player, int cash) {
         User user = Township.getUserManager().getUser(player.getUniqueId());
         user.setCash(cash);
         player.sendMessage(Msg.format("<green>Set cash to %d!", cash));
